Now works as intended. Very slow though
The slowness is due to the freeipa user-add line. One fix for this is to
go for multithreading. But this is a matter for another time. Something
that should also be changed is the call to os.system. A solution like
subprocess.Popen should progrably used instead. But for now, fuck that,
it works.